in process: alterin app linkages between actual ifferent screns, namely: Settomgs.
home/map
login/register

additional link to testing console
