+++ app/src/main/java/com/example/wildfire_fixed_imports/MainActivity.kt
@@ -37,7 +37,7 @@ class MainActivity : AppCompatActivity() {
         appBarConfiguration = AppBarConfiguration(
             setOf(
                 R.id.nav_home, R.id.nav_login_register, R.id.nav_settings,
-                R.id.nav_tools, R.id.nav_share, R.id.nav_send
+                R.id.nav_debug, R.id.nav_share, R.id.nav_send
             ), drawerLayout
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
