+++ app/src/main/java/com/example/wildfire_fixed_imports/view/home/HomeFragment.kt
@@ -1,13 +1,10 @@
-package com.example.wildfire_fixed_imports.ui.home
+package com.example.wildfire_fixed_imports.view.home
 
-import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.example.wildfire_fixed_imports.R
 import com.mapbox.mapboxsdk.Mapbox
