+++ app/src/main/java/com/example/wildfire_fixed_imports/view/home/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.wildfire_fixed_imports.ui.share
+package com.example.wildfire_fixed_imports.view.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ShareViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is share Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
