+++ app/src/main/java/com/example/wildfire_fixed_imports/view/login_register/GalleryViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.wildfire_fixed_imports.ui.home
+package com.example.wildfire_fixed_imports.view.login_register
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class GalleryViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
