+++ app/src/main/res/values/strings.xml
@@ -10,7 +10,7 @@
     <string name="menu_home">Home</string>
     <string name="menu_login_register">Login</string>
     <string name="menu_settings">Settings</string>
-    <string name="menu_tools">Tools</string>
+    <string name="menu_debug">Debug</string>
     <string name="menu_share">Share</string>
     <string name="menu_send">Send</string>
